+++ test/wallaby/selenium/webdriver_client_test.exs
@@ -1,7 +1,7 @@
-defmodule Wallaby.Experimental.Selenium.WebdriverClientTest do
+defmodule Wallaby.WebdriverClientTest do
   use Wallaby.HttpClientCase, async: true
 
-  alias Wallaby.Experimental.Selenium.WebdriverClient, as: Client
+  alias Wallaby.WebdriverClient, as: Client
   alias Wallaby.{Element, Query, Session}
 
   @web_element_identifier "element-6066-11e4-a52e-4f735466cecf"
@@ -81,7 +81,7 @@ defmodule Wallaby.Experimental.Selenium.WebdriverClientTest do
                parent: session,
                session_url: session.url,
                url: "#{session.url}/element/#{element_id}",
-               driver: Wallaby.Experimental.Selenium
+               driver: Wallaby.Selenium
              }
     end
 
@@ -114,7 +114,7 @@ defmodule Wallaby.Experimental.Selenium.WebdriverClientTest do
                parent: parent_element,
                session_url: session.url,
                url: "#{session.url}/element/#{element_id}",
-               driver: Wallaby.Experimental.Selenium
+               driver: Wallaby.Selenium
              }
     end
 
@@ -141,7 +141,7 @@ defmodule Wallaby.Experimental.Selenium.WebdriverClientTest do
                parent: session,
                session_url: session.url,
                url: "#{session.url}/element/#{element_id}",
-               driver: Wallaby.Experimental.Selenium
+               driver: Wallaby.Selenium
              }
     end
   end
@@ -1111,7 +1111,7 @@ defmodule Wallaby.Experimental.Selenium.WebdriverClientTest do
     session_url = bypass_url(bypass, "/session/#{session_id}")
 
     %Session{
-      driver: Wallaby.Experimental.Selenium,
+      driver: Wallaby.Selenium,
       id: session_id,
       session_url: session_url,
       url: session_url
@@ -1120,7 +1120,7 @@ defmodule Wallaby.Experimental.Selenium.WebdriverClientTest do
 
   defp build_element_for_session(session, element_id \\ ":wdc:abc123") do
     %Element{
-      driver: Wallaby.Experimental.Selenium,
+      driver: Wallaby.Selenium,
       id: element_id,
       parent: session,
       session_url: session.url,